Fix fonts not being embedded and remove default font
I previously noticed that the fonts were not rendering for a couple of projects, and the font happened to be Sans Serif, so I made the weird conclusion that Scratch must have a "default" font-family applied to all text, that font being Sans Serif.

But this was not the case. Instead, it appeared my method of detecting the font-family attribute when embedding the fonts was wrong. I assumed that only <text> elements could have the font-family attribute, but it turns out that any other elements such as <g> can also have it, which applies the font to descendant <text> elements, as was the case in the projects where the font(s) wouldn't load. Whoops!

So I have revised the code to check ALL the elements of the <svg> to see if they have the font-family attribute set and if so, embed the preset font file for that font-family to render (if available). The default font has been removed as well, since the problem it was added to patch has now been more properly fixed by the aforementioned change.

😮‍💨